Refactored the code of Migration
